Better support of entities and '&'
The "entity" renderer callback is no longer called when a lone '&' is
encountered, only when a valid entity ("&#?[A-Za-z0-9]+;") is recognized.
Otherwise '&' is considered as an inactive character and handed over to
"normal_text" callback along with other inactive charaters.

In the example renderers rndr_entity has therefore been replaced by NULL,
because the only treatment of entities is to copy them verbatim into the
output (without the escaping rndr_normal_text does).