markdown: add missing braces in an `if` block
The second `if` should be inside the first, as denoted by the indentation.
Fortunately, this doesn't result in a visible bug, because `j` is
initialized to zero and not touched otherwise.
However, it does trigger a warning in gcc-6, which causes a build failure
because of `-Werror`, reported in https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=811587

Thanks to Massimo Manghi for letting me know.