buffer: remove "already-done" check in `bufnullterm`
The point of this check was to do nothing when the first byte after
used data exists and is already 0. So it saves a memory write (and
possible cache invalidation) at the cost of a comparison. The trade-off
is far from obvious.

On top of that, it is counted as a conditional that depends on
uninitialized memory by valgrind (as reported in #15 ), which is
technically true, even though that particular instance is harmless.

So I decided to remove this check and unconditionally write 0 after
the end of the current data (after making for it if needed).